added new file and test file for SLSimObject.
